+++ fixture/package-config-2/metalsmith.js
@@ -1,0 +1,13 @@
+/* istanbul ignore next */
+(function() {
+  var docpress = require('../../ms')
+
+  var app = docpress(__dirname)
+    .use(require('../../')())
+
+  if (module.parent) {
+    module.exports = app
+  } else {
+    app.build(function (err) { if (err) throw err })
+  }
+}())