+++ HISTORY.md
@@ -4,6 +4,7 @@
 - Markdown code is now stripped out of TOC names
 - The first page will always be the home page
 - You can now link Table of Contents as part of the site
+- Make heading ID's match what GitHub expects
 
 ## [v0.2.0]
 > Oct 12, 2015