+++ HISTORY.md
@@ -1,4 +1,5 @@
-## master
+## [v0.3.0]
+> Oct 14, 2015
 
 - Table of contents (`docs/README.md`) is now optional
 - Markdown code is now stripped out of TOC names
@@ -22,3 +23,4 @@
 - Initial release.
 
 [v0.2.0]: https://github.com/docpress/docpress-core/compare/v0.1.0...v0.2.0
+[v0.3.0]: https://github.com/docpress/docpress-core/compare/v0.2.0...v0.3.0